Accept 64-it integers in ScaledLinearHistogram::AddScaledCount()
Values passed to AddScaledCount() can be quite big, and we need them to
fit into int32 only after scaling.

Bug: 1407438
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191808
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/main@{#1096699}

+++ base/metrics/histogram.h
@@ -435,7 +435,8 @@ class BASE_EXPORT ScaledLinearHistogram {
   // Like AddCount() but actually accumulates |count|/|scale| and increments
   // the accumulated remainder by |count|%|scale|. An additional increment
   // is done when the remainder has grown sufficiently large.
-  void AddScaledCount(Sample value, int count);
+  // The value after scaling must fit into 32-bit signed integer.
+  void AddScaledCount(Sample value, int64_t count);
 
   int32_t scale() const { return scale_; }
   HistogramBase* histogram() { return histogram_; }
